[System] Port System.Net.Mime namespace from referencesources
A few helper methods that existed in ContentType were moved into MailMessage.
I verified the test changes are required to match MS.NET behavior.
