+++ mcs/class/System/ReferenceSources/Logging.cs
@@ -29,6 +29,9 @@ namespace System.Net {
 		internal static void Exit(TraceSource traceSource, object obj, string method, object retObject) {
 		}
 
+		internal static void PrintInfo(TraceSource traceSource, object obj, string method, string msg) {
+		}
+
 		internal static void PrintWarning(TraceSource traceSource, object obj, string method, string msg) {
 		}
 	}
