--- mcs/class/System/System.Net.Mime/ChangeLog
@@ -1,90 +1,0 @@
-2009-09-23  Zoltan Varga  <vargaz@gmail.com>
-
-	* ContentType.cs: Fix the net 2.1 build.
-
-2009-04-02  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : when there are especials, quote parameter values.
-
-2008-08-14  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : patch by Ted Unangst, fixed bug #392843.
-	  handle RFC2047 encoding.
-
-2008-02-16  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : Reject such MediaType that lacks '/'.
-	  Do not reject such parameters that are not standard ones.
-
-2007-12-05  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : for SubjectEncoding, null is returned instead of 
-	  ASCII (.net compat way). For this cosmetic change, BodyEncoding is 
-	  changed to fill ASCII automatically.
-
-2007-12-05  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : added GuessTransferEncoding() to guess correct
-	  encoding type for sys.net.mail.Attachment.
-	  Here UTF8Encoding should not include BOM.
-
-2007-10-22  Atsushi Enomoto  <atsushi@ximian.com>
-
-	* ContentType.cs : moved couple of encoding stuff to here.
-	  Encode non-ASCII value in RFC 2047.
-
-2007-09-28  Marek Habersack  <mhabersack@novell.com>
-
-	* ContentDisposition.cs: if the file name contains spaces, enclose
-	it in quotes. Fixes bug #324084.
-
-2006-01-13  John Luke  <john.luke@gmail.com>
-
-	* ContentDisposition.cs: use String.Compare instead
-	of comparing .ToLower's
-	
-2006-01-02  John Luke  <john.luke@gmail.com>
-
-	* ContentDisposition.cs: remove [MonoTODO]'s
-	access and store the properties in Parameters
-	ToString is now just DispositionType + the parameters
-	* ContentType.cs: remove [MonoTODO]'s
-	use parameters to store Name, Boundary, and CharSet
-	ToString is now just MediaType + the parameters
-	throw FormatException for unrecognized paremeter in ctor
-	* TransferEncoding.cs: remove [Serializable]
-
-2005-12-26  John Luke  <john.luke@gmail.com>
-
-	* ContentDisposition.cs: Size is -1 by default
-	throw ArgumentNullException and ArgumentException for
-	set_DispositionType
-	* ContentType.cs: throw ArgumentNullException, ArgumentException,
-	and FormatException for set_MediaType
-	don't print Boundary,Name in ToString() if they are ""
-
-2005-12-23  John Luke  <john.luke@gmail.com>
-
-	* ContentDisposition.cs: override GetHashCode and
-	Equals, implement ToString,
-	fix Size to be long not int,
-	implement Inline property,
-	implement the ctors
-	* ContentType.cs: override GetHashCode and Equals
-	implement ToString, implement the ctors
-	
-2005-12-22  John Luke  <john.luke@gmail.com>
-
-	minor updates I noticed for 2.0 final API
-	* DispositionTypeNames.cs: change to a static class
-	* MediaTypeNames.cs: change these to static classes
-	* TransferEncoding.cs: remove Binary and EightBit
-	add [Serializable] and fix the values
-	
-2004-09-09  Tim Coleman <tim@timcoleman.com>
-	* ContentType.cs: Changes to MIME headers
-
-2004-09-04 Tim Coleman  <tim@timcoleman.com>
-	* ContentDisposition.cs ContentType.cs DispositionTypeNames.cs
-	* MediaTypeNames.cs TransferEncoding.cs:
-		New class stubs for 2.0
