+++ mcs/class/System/System.dll.sources
@@ -337,11 +337,36 @@ System.Net.Mail/SmtpFailedRecipientsException.cs
 System.Net.Mail/SmtpPermission.cs
 System.Net.Mail/SmtpPermissionAttribute.cs
 System.Net.Mail/SmtpStatusCode.cs
-System.Net.Mime/ContentDisposition.cs
-System.Net.Mime/ContentType.cs
-System.Net.Mime/DispositionTypeNames.cs
-System.Net.Mime/MediaTypeNames.cs
-System.Net.Mime/TransferEncoding.cs
+../../../external/referencesource/System/net/System/Net/mail/Base64Stream.cs
+../../../external/referencesource/System/net/System/Net/mail/Base64WriteStateInfo.cs
+../../../external/referencesource/System/net/System/Net/mail/BaseWriter.cs
+../../../external/referencesource/System/net/System/Net/mail/BufferBuilder.cs
+../../../external/referencesource/System/net/System/Net/mail/ClosableStream.cs
+../../../external/referencesource/System/net/System/Net/mail/ContentDisposition.cs
+../../../external/referencesource/System/net/System/Net/mail/ContentType.cs
+../../../external/referencesource/System/net/System/Net/mail/DelegatedStream.cs
+../../../external/referencesource/System/net/System/Net/mail/DispositionTypeNames.cs
+../../../external/referencesource/System/net/System/Net/mail/EightBitStream.cs
+../../../external/referencesource/System/net/System/Net/mail/EncodedStreamFactory.cs
+../../../external/referencesource/System/net/System/Net/mail/HeaderCollection.cs
+../../../external/referencesource/System/net/System/Net/mail/IEncodableStream.cs
+../../../external/referencesource/System/net/System/Net/mail/MailBnfHelper.cs
+../../../external/referencesource/System/net/System/Net/mail/MailHeaderInfo.cs
+../../../external/referencesource/System/net/System/Net/mail/MimeBasePart.cs
+../../../external/referencesource/System/net/System/Net/mail/MimeEmbeddedMessagePart.cs
+../../../external/referencesource/System/net/System/Net/mail/MimeMultiPart.cs
+../../../external/referencesource/System/net/System/Net/mail/MimeMultiPartType.cs
+../../../external/referencesource/System/net/System/Net/mail/MimePart.cs
+../../../external/referencesource/System/net/System/Net/mail/MimeWriter.cs
+../../../external/referencesource/System/net/System/Net/mail/MultiAsyncResult.cs
+../../../external/referencesource/System/net/System/Net/mail/QEncodedStream.cs
+../../../external/referencesource/System/net/System/Net/mail/QuotedPrintableStream.cs
+../../../external/referencesource/System/net/System/Net/mail/SmtpDateTime.cs
+../../../external/referencesource/System/net/System/Net/mail/TrackingStringDictionary.cs
+../../../external/referencesource/System/net/System/Net/mail/TrackingValidationObjectDictionary.cs
+../../../external/referencesource/System/net/System/Net/mail/TransferEncoding.cs
+../../../external/referencesource/System/net/System/Net/mail/WriteStateInfoBase.cs
+../../../external/referencesource/System/net/System/Net/mail/mediatypenames.cs
 System.Net/MacProxy.cs
 System.Net/MonoHttpDate.cs
 System.Net/NetConfig.cs
