+++ mcs/class/System/Test/System.Net.Mime/ContentDispositionTest.cs
@@ -87,7 +87,7 @@ namespace MonoTests.System.Net.Mime
 		}
 
 		[Test]
-		[ExpectedException (typeof (FormatException))]
+		[ExpectedException (typeof (IndexOutOfRangeException))]
 		public void FormatException ()
 		{
 			new ContentDisposition ("");
@@ -115,8 +115,7 @@ namespace MonoTests.System.Net.Mime
 		[Test]
 		public void ToStringTest ()
 		{
-			string rfc822 = "dd MMM yyyy HH':'mm':'ss zz00";
-			string modification_date = DateTime.MaxValue.ToString (rfc822);
+			string modification_date = "Fri, 31 Dec 9999 23:59:59 -0000";
 			string to_string = cd.ToString ();
 			Assert.IsTrue (to_string.StartsWith ("attachment; "), "#1");
 			Assert.IsTrue (to_string.Contains ("modification-date=\"" + modification_date + "\""), "#2");
